2014 MLE

+++ acmicpc/src/2014.cpp
@@ -16,13 +16,14 @@ int main(void)
         s.insert((ll)A[i]);
     }
     ll t;
+
     for (int i = 0; i < N; i++)
     {
         t = *s.begin();
         s.erase(t);
 
         for (int j = 0; j < K; j++)
-            if (t * A[j] < (1 << 31))
+            if (t * A[j] < ((ll)1 << 31))
                 s.insert(t * A[j]);
     }
     printf("%lld\n", t);
